Storybook: Line - Simplify API story and extract Examples
