+++ storybook/stories/API/cartesian/Line.mdx
@@ -1,7 +1,13 @@
-import { ArgTypes, Meta, Stories } from '@storybook/blocks';
+import { Meta, Story, Canvas, Controls } from '@storybook/blocks';
 import * as ComponentStories from './Line.stories';
 
 # Line
+
+<Meta of={ComponentStories} />
+
+<Canvas>
+  <Story name="API" of={ComponentStories} />
+</Canvas>
 
 ## Parent Component
 The Line can be used within a `<ComposedChart/>` or a `<LineChart/>`.
@@ -13,9 +19,7 @@ The Line can be used with the following child components: `<ErrorBar/>` `<LabelL
 
 Properties in the groups Other and Internal are not recommended to be used.
 
-<Meta of={ComponentStories} />;
-<ArgTypes of={ComponentStories} sort={'requiredFirst'} />;
-<Stories />;
+<Controls of={ComponentStories.API} />
